hkutils.configparser_to_configdict: added
[config file]

hkutils.configparser_to_configdict is a new function that converts ConfigParser
configuration objects into dictionaries. These dictionaries can be handled more
easily, especially when there are subsection and subsubsections.

For example, the following config file will be transformed to the following
dictionary (this is *not* a real Heapkeeper config file, the new Heapkeeper
config file format that supports several heaps is not yet clear):

    [paths]
    html_dir=/home/me/html

    [heaps/hh]
    heap_id=hh
    name=Heapkeeper heap
    path=/home/me/hh

    [heaps/personal]
    heap_id=my
    name=Personal heap
    path=/home/me/personal

The following dictionary will be returned by configparser_to_configdict:

    {'paths': {'html_dir': '/home/me/html'}},
     'heaps': {'hh': {'heap_id': 'hh',
                      'name': 'Heapkeeper heap',
                      'path': '/home/me/hh'},
               'personal': {'heap_id': 'my',
                            'name': 'Personal heap',
                            'path': '/home/me/personal'}}

This is easier to use then the ConfigParser object, e.g.:

- List of heaps:

    d['heaps'].keys()

- Paths to the heaps:

    [heap['path'] for heap in d['heaps'].values()]

- Any subtree of the configuration can be passed as an argument.