hkbodyparser.Segment: type=meta -> is_meta=True
[post body parsing]

Segments have a different method to store that they are meta than previously.
Previously their `type` attribute could be 'meta'. Now they have a new
attribute called `is_meta` which is True or False. The reason for this change
is that now meta segments can be links, heap links, or can have any other
segment type. This makes it possible to include links in meta text for example.