hkbodyparser: parsing raw blocks
[post body parsing]

Raw blocks are parsed in post bodies.

A raw block is defined in the following way:
- Each non-empty line of a raw block starts with a space or a tab (not counting
  quote signs).
- If there is a line before the raw block, it should be either empty (not
  counting quote signs) or be on a different quote level.
- If there is a line after the raw block, it should be either empty (not
  counting quote signs) or be on a different quote level.

I found only one problematic point of this definition. The following example
will be parsed incorrectly:

    1. Let's look at x:

            x = 0

       This example sets x to 0.

Heapkeeper will consider the last line of the example as part of the raw block.
An idea was to say that all non-empty lines of the raw block should have at
least as many spaces or tabs in their beginning as the first line of the block;
but that would not work because that is usually not true for diff blocks:

     Something unchanged
    -Something removed
    +Something added
     Something unchanged

Raw blocks are shown in gray in the generated HTML pages.