hkshell.enew: adds date to the post
[hkshell]

hkshell.enew() can now add a date to the post that is created. The new
parameter `dt` has been added to hkshell.enew(). This parameter specifies what
date to add. By default, `dl` is None, in which case `enew` will add the
current date. The special value 0 can be used to specify that no date should be
added to the post.