Generator: print_html_header_info is deprecated
[generator]

The print_html_header_info method printed the date on which the HTML page was
generated and the Heapkeeper version which was used. At first this seemed to be
a good idea.

But it turned out to be a problem in the following cases:
- I have a script that compares the HTML files generated by different
  Heapkeeper versions so that I can check what has changed in the Generator.
- The HTML files for the Heapkeeper Heap are stored in a GitHub repository.
  Since all HTML files are regenerated before every commit to that repository,
  all files changed in all commit because of this header info.

These scripts turned off printing the header info to avoid these problems.

I now decided to remove the header info, since I have seen no benefits
whatsoever, but I have seen these two problems which could effect anybody.