hkweb: folding into a button
[hkweb]

Buttons have been added to the top of post body: folding can now be initiated
with a "Hide" button instead of clicking on the body.

Previously the "folding" feature folded the post body into a one-line box that
replaced the body. Now the body is totally hidden when the user folds it, and a
"Show body" button appears in the post summary area that can be used to re-open
it.