hklib: maximum IMAP command length reduced
[imap bugfixes]

IMAP code has undergone a major rewrite to shorten the length of IMAP
command lines. This is done by using IMAP sequence notations whenever
possible, and breaking commands that would be too long into multiple
shorter commands. RFC 2683 requires clients to limit their commands to
a maximum of 1000 octets.

Note that in practice, we will almost never need to fragment a command
to respect this rule thanks to the IMAP sequence notation, but the
code is there just to be safe.

This corrects heap://hh/1211.