hkweb: post bodies can be edited
[hkweb]

Edit functionality has been added, using AJAX. Instead of XML, we use JSON as a
common message format between the client and the server.

CSS classes and ids were modified to contain more hyphens instead of underscores and
slashes. For example the CSS id for the summary node of the post my_heap/0 was
"post_my_heap/0", but now it is "post-my_heap-0". Underscores are abandoned
because they do not express that they separate two parts of the id. Slashes are
abandoned because they have a special meaning in the JQuery selectors.