hkweb: newlines added after HTML elements
[hkweb]

Previously there were some HTML elements that had no whitespace between them.
This caused problems. For example if the following buttons are present on a
page:

    <span>Button One</span><span>Other Button</span>

then the browser may break the line between the words "Button" and "One", and
between "Other" and "Button", but it will not break it between "One" and
"Other", because there is no whitespace there. To solve this problem, I
appended a newline character at the end of each problematic HTML elements:

    <span>Button One</span>
    <span>Other Button</span>

This also has a nice side effect that the generated HTML code is easier to
read.