hkshell: ImportError shown when appropriate
[bugfix]

When hkshell imports a module specified by a user, and the module is not found
(thus an ImportError is raised), hkshell discards the exception with the stack
trace, and prints only a one line message that the module was not found.

The problem with the previous implementation was that it discarded the
exception in case of all ImportErrors, not only ImportErrors that belonged to
the module specified by a user. So for example if a user imported
mysettings.py, which imported web.py, and web.py was not found, the error
message was: 'Module not found: mysettings'.

Now ImportErrors is discarded only if it belongs to the module specified by the
user, otherwise it is re-raised. So in the example, the exception about not
finding web.py would be printed on the screen.