Post.parse_header: space after colon in optional
[bugfix]

Previously, header attributes in post files had to have a space after the colon
to separate the key and the value of the attribute, for example:

    Subject:<space>Funny

Now this space is made optional. This has the effect that the following header
attributes (=lines in the header of a post file) are valid:

    Subject:
    Subject:Funny

I mostly committed this fix for the first one; it was wrong to require a
trailing space at the end of the "Subject:" line. But as a side effect of this
fix, the second one became valid as well; which does not mean that that form
should be used.