hkp_review: replacing method instead of class
[hkweb]

It turned out to be a bad idea to replace hkweb.PostPageGenerator with another
(similar) class. The problem became apparent when I wanted to subclass
PostPageGenerator. Let's call the subclass SpecificPostPageGenerator. So the
following is the problematic code:

    hkweb.py:

        class PostPageGenerator(...):
            ...

        class SpecificPostPageGenerator(PostPageGenerator):

            def __init__(self, ...):
                PostPageGenerator.__init__(self, ...)   ###
                ...
            ...

    hkp_review.py:

        OldPostPageGenerator = hkweb.PostPageGenerator
        class PostPageGenerator(OldPostPageGenerator):
            ...
        hkweb.PostPageGenerator = PostPageGenerator

So there are two versions of the PostPageGenerator class: the original one,
defined in hkweb, and the new one that is defined in hkp_review and which wraps
the original one.

When an instance of SpecificPostPageGenerator is created, the line noted with
"###" will throw an exception. The reason is that SpecificPostPageGenerator is
subclassed from the original PostPageGenerator (because at the moment of the
definition of SpecificPostPageGenerator, hkweb.PostPageGenerator is set to the
original PostPageGenerator), but the line "###" will call the new
PostPageGenerator's __init__ method (because at the moment of the function
call, hkweb.PostPageGenerator is set to the new PostPageGenerator).

This particular problem could be solved probably by using the `super` function,
but that would not be perfect, because then the modifications done by the
hkp_review module would not affect the instances of SpecificPostPageGenerator.
(I don't know whether we want to affect the instances like these in the case of
other plugins, but here that would be more logical.)

So I solved the problem in this commit by not creating another class instead of
PostPageGenerator, only creating a new __init__ method for it (because that
is the method whose behaviour the plugin wanted to change).