hkweb.js.gotoURL: added
[hkweb]

The ajaxQuery function does a good job in converting a JavaScript object into
JSON-coded HTTP query parameters, and invoking a callback function with the
server's response.

I also needed a similar function that would load a given URL. This function
(gotoURL) also converts the given JavaScript object into JSON-coded HTTP query
parameters, but afterwards it loads the URL in the browser with these
parameters instead of processing it.

JQuery helped in writing ajaxQuery because it converts the given JavaScript
object (the `data` variable) into HTTP query parameters. I could not find such
help when writing gotoURL, so I had to code up that conversion, which takes up
all but one lines of the gotoURL function.