hkweb: search functionality added
[hkweb]

The URL http://localhost:8080/search shows a search bar. If the user types
words and presses "Search the heaps", a page with the results will be shown.
The posts that contained all words (which are in fact regular expressions
matched against the body and the header attributes of the post) will be shown
as active posts, and their thread mates will be shown as passive.