hkweb: string query params accepted, json escaped
[hkweb]

This commit makes it possible to accept plain strings as HTTP query parameters
by escaping the JSON query parameters.

== Explanation ==

Up until commit 817322677a, all parameters in a hkweb HTTP query were passed in
one query parameter as one JSON object, e.g.:

    http://localhost:8080/search?args={"term":"hello%20there"}

(The searching functionality did not exist then, but if it had, this is what it
would have looked like.)

After commit 817322677a, each parameter became one query parameter, but as a
stringified JSON object:

    http://localhost:8080/search?term="hello%20there"

(Note the double quotes in the URL.)

Now strings are passed as strings and not as JSON objects:

    http://localhost:8080/search?term=hello%20there

We would still like to be able to pass JSON objects. When we want to do that,
we have to escape it with a zero character:

    http://localhost:8080/search?posts=%00["hh/1","hh/2"]

We can use the JSON notation even when sending a string. The following line is
as same as if we wrote `term=hello%20there`:

    http://localhost:8080/search?term=%00"hello%20there"

== Reason for the change ==

The reason for this change is that the search functionality in hkweb could not
have been used previously using HTML forms, only using JavaScript, because we
cannot tell the HTML forms to convert the search expression typed by the user
into a JSON string. I tried to put a search bar to the index page and write
JavaScript code to initiate the search, but it turned out to be difficult. I
decided that it is not OK that it was that hard, so I committed this change
which makes HTML forms work. They now can just send the values of their input
fields without modification.