hkweb{,.js}: adding new posts implemented
[hkweb]

The modification affects both the server and the client side.
Modification of the server side is minimal: besides adding the button,
the SetPostBody server has been extended to handle adding a new post
as a child to an existing post. This would probably be more elegantly
handled as a new server class that does nothing else.

The JS side is more complicated: it copies the summary of the parent
post on clicking the "Add child" button, then modifies the clone as
needed. When the user clicks "Save", the content is sent to the server,
which sends back the post summary and the post id of the new post.
This is used to display the resulting post and to bind event handlers
to its buttons.