hkshell: fixed bug in importing modules
[bugfix]

The mechanism that checks whether the default hkrc module is present
before importing it was broken: it checked in the working dir, and
imported from the Python path.

The new mechanism doesn't check if the file is there; instead it
suppresses any import warnings if the default hkrc is not present.