hkweb: HTTP basic authentication added
[hkweb auth]

This is the first step towards a hkweb with true authentication. HTTP
authentication is working, but the username and password are hardcoded
into hkweb.

Note: the word 'basic' denotes the authentication method used, and is
not a reference to the implementation being rudimentary. This method
involves the username/password pair being transmitted in plaintext.
Always keep in mind the security consequences of this.