hksearch: adding more target types (e.g. 'heap')
[search]

When the users search in the post database, they enter a search term, which is
a list of targets. Only the posts that match all targets will be shown. So far
the only target type was the regexp that was searched for in the whole post. So
for example if the search term was "abc", all post attributes and the body was
searched for "abc", and if any of them contained it, the post was considered as
a match.

In this commit, new target types have been introduced. The target type that was
previously the only one is now called 'whole' (because it examines the whole
post, including all attributes and the body). The new ones are the following:
'heap', 'author', 'subject', 'tag', 'message-id', 'body'. If a search target is
e.g. 'heap:myheap', it will match only posts whose heap contains the regular
expression 'myheap'. The other target types are similar.

Examples for search terms that use the new target types:

- "heap:^hh$ release" -- Search for the "release" regexp in the posts of the 'hh'
  heap. "^" means "beginning of string" and "$" means "end of string" in the
  regexp language.
- "tag:python|C++ \befficiency\b" -- Search for the word "efficiency" in posts
  that have "python" or "C++" tags (or rather in posts that have a tag that
  contains the regexp "python" or "c++"). \b means "word boundary" in the
  regexp language.

The search engine could be further improved. A few possibilities:
- Currently we are always case insensitive. The user could turn that off.
- Currently we always use regular expressions. The user could turn that off and
  have Heapkeeper interpret the search strings literally.
- We could use double quotes to capture spaces like Google does. (Currently \s
  can be used to include a space character in a search target.)
- Negation: e.g. "-tag:python" could mean "posts that do not have a "python"
  tag. This is also done by Google.