hklib: added accounts to config file format 3
[hkweb auth]

The config file now stores usernames and passwords for user accounts.
These will be used in the final stage of implementing HTTP
authentication.

The account entries are stored in a separate section of the config
file, which looks like this:

    [accounts]
    adam=apple
    eve=snake

The part before the equation sign is the username, the part after it
is the password.