hkp_users: added plugin to show active users
[users plugin]

The output is simplistic, but tries to be as nonobtrusive as possible.
E.g. the time is hidden behind the name, and visible on mouseover
only.

The box now follows the view (in CSS, its position is fixed), but it
may be even less obtrusive if it were anchored to the document top
(with absolute positioning).

Users time out of the view after 5 minutes of inactivity.