hkp_chat: added separate channels for threads
[chat plugin]

Now users see each others' lines only if they are in the same
"channel", that is, they are viewing the same thread in hkweb. There
is also a separate "global" channel for the index.

TODO: later we could add a way to switch to another channel without
changing views. Also, the user should have a way to check recent
activity on any channel.