doc/Makefile: strict target added
[doc]

A new target called "strict" was added to the Makefile that builds the
documentation. This does the same as the "html" target (i.e. it compiles all
Sphinx documentations and generates HTML documentation), but it asks Sphinx to
convert the warnings into errors. This is useful for hk-dev-utils which wants
to have "make" to have a non-zero return code in case of any warning or error.