hkweb: authentication keeps track of last access
[hkweb auth]

The hkweb module now keeps the time of the last access in a dict. This
makes it possible to check if anyone is using the service
through the Web interface before making any drastic changes on the
command line.

This use is further supported by the convenience function
`hkweb.last`:

    >>> hkweb.last()
    Last access was 42 seconds ago.

This feature only works if authentication is enabled.