hkgen.Generator: split into two classes
[refactoring]

The hkgen.Generator class has been split into two classes:
- hkgen.BaseGenerator, which does not contain functions specific to static HTML
  generation;
- hkgen.StaticGenerator, which contains functions only specific to static HTML
  generation.

We talked about removing hkcustomlib in heap://hh/1282 and heap://hh/1295.

First I wanted to call the BaseGenerator simply Generator, but that would have
made it more difficult to make hkgen.Generator deprecated. Users may use
hkgen.Generator in their hkrc files, so it could not simply lack all
static-HTML-specific functions. If it had those functions, that would have
caused code duplication in the code of this new hkgen.Generator and
hkgen.StaticGenerator. So I created a separate hkgen.Generator class that is
deprecated and will be removed. Also, it may be good that developers see that
it is not a complete generator but it needs to be subclassed to make it usable
by adding the capability of writing the results to files or serving them as
response to web requests.