hkgen: escape_link renamed to escape_url
[refactoring]

From heap://hh/1337:

    Some function names, variable names and other identifiers in hkgen.py
    contain the word "link", but they actually mean "URL". These identifiers
    should be modify to contain the word "url".

I have found only this function in hkgen with this problem. Everywhere else in
hkgen the word "link" means an actual link.

We could rename the "links" to "urls" in hkbodyparser, but I'm not sure that is
a good idea. There "link" means "a URL that should be shown as a link", and
later we may include text here which is not a URL but still should be shown as
a link.