hkweb.AjaxServer: not accepting GET requests
[hkweb]

hkweb.AjaxServer does not accept GET requests any more, only POST requests. The
reason is that it is easier for malicious people to take advantage of GET
requests that perform some change in the Heapkeeper database. Consider that
someone sends an email like this in an email:

    Check this out: http://hh.com/set-post-body/hh/1?new_body_text=XXX

If someone would have clicked on that link, the browser would have sent the
request to set the body of the post by just following it. With POST requests,
this is not so easy.