hk_issue_tracker: moved to hkp_issue_tracker
[hkweb cleanup]

Previously, the issue tracker was broken into two parts: it was partly in
src/hk_issue_tracker.py, and partly in
plugins/issue_tracker/src/hkp_issue_tracker.py. Now it is completely in the
latter, which means that the issue tracker is not part of Heapkeeper's core.

Both the static and web output work, and both the flat and threaded output
work.