hkweb: print_additional_* funs have 'info' param
[hkweb cleanup]

The print_additional_header and print_additional_footer method of
hkweb.WebGenerator have been refactored. Now instead of having different
parametrization in different subclasses (e.g. PostPageGenerator had an extra
parameter called `postid`), the parameters are always the same: `self` and
`info`. The latter one is a dictionary that contains the necessary parameters,
e.g. in case of PostPageGenerator, it contains a 'postid' key.

This way, it will be much easier to introduce new parameters, because most
function will just pass `info` along without having to care about its content.
Look at for example the modifications done to the
hkp_users.postpage_print_additional_header function:

    -def postpage_print_additional_header(self, postid):
    -    return (postpage_old_print_additional_header(self, postid),
    +def postpage_print_additional_header(self, info):
    +    return (postpage_old_print_additional_header(self, info),
                 print_userlist(self))

The new version of the function does not have to care about `info`, that will
just be passed along to the original postpage_old_print_additional_header
function, so it is not specific to PostPageGenerator.

When the value of a parameter is needed, of course it is read from `info`, just
like here:

    -def postpage_print_additional_header(self, postid):
    -    return (postpage_old_print_additional_header(self, postid),
    -            threadid_js(postid),
    +def postpage_print_additional_header(self, info):
    +    return (postpage_old_print_additional_header(self, info),
    +            threadid_js(info['postid']),
                 print_chat(self))

Also, since having an `info` argument is more general, we will be able to write
some functions in a more generic way. E.g. now the
index_print_additional_header and postpage_print_additional_header functions in
hkp_users are the same, so the code can be refactored to have only one of them.