hkweb: favicon.ico is served
[hkweb cleanup]

If the location of the favicon is not specified explicitly, Google Chrome will
look for /favicon.ico. This occurs for example when an error happens and
Heapkeeper serves only an error string as response, and when someone looks at
the source code of a page. So now Heapkeeper serves /favicon.ico.