hkutils: hk_dir variable added
[hkweb cleanup]

The hkutils.hk_dir variable have been created which contains the path to the
main Heapkeeper directory. This is a step into the direction of Heapkeeper not
requiring to be started from its main directory (see heap://hh/1337).

This became necessary for the test of hkweb.Fetch, because the test case
changed the current directory to the temporary directory created for it so that
it can work with the config file with relative paths; but this caused
Heapkeeper not to find the static/ directory.