hkshell: invoking hkrc.start
[hkweb cleanup]

hkshell now executes the start() function of the hkrc files if they exist. This
is necessary because hkrc files are not allowed to run hkweb.start in their
main body, since that leads to a deadlock.

The hkrc_start_invoked variable has been added to hkshell so that hkrc modules
can be written to check whether this variable exists and call their start()
function themselves if it does not.